+++ frontend/src/pages/HomePage.tsx
@@ -1,53 +1,26 @@
-import { Box, Button, Container, Stack, Typography } from "@mui/material";
+import { Box } from "@mui/material";
 import type { ReactElement } from "react";
-import { useNavigate } from "react-router-dom";
 
-import KapitelShelfLogo from "/kapitelshelf.png";
-
-import FancyText from "../components/FancyText";
-import { useMobile } from "../hooks/useMobile";
+import WelcomeScreen from "../features/WelcomeScreen";
+import { useWelcomeScreen } from "../hooks/useWelcomeScreen";
 
 const HomePage = (): ReactElement => {
-  const navigate = useNavigate();
-  const { isMobile } = useMobile();
+  const [isWelcomeShown] = useWelcomeScreen();
 
-  return (
-    <Box
-      sx={{
-        mt: "12%",
-        display: "flex",
-      }}
-    >
-      <Container sx={{ textAlign: "center" }}>
-        <Stack
-          direction="row"
-          justifyContent="center"
-          alignItems="center"
-          spacing={2}
-        >
-          <img
-            src={KapitelShelfLogo}
-            alt="KapitelShelf logo"
-            style={{ width: isMobile ? "70px" : "100px" }}
-          />
-          <FancyText variant="h2">KapitelShelf</FancyText>
-        </Stack>
-        <Typography variant="h6" color="text.secondary" mt="15px">
-          Welcome to your personal library. Manage your books, track your
-          reading, and explore your collection.
-        </Typography>
-        <Box mt={4}>
-          <Button
-            variant="contained"
-            size="large"
-            onClick={() => navigate("/library")}
-          >
-            View My Books
-          </Button>
-        </Box>
-      </Container>
-    </Box>
-  );
+  if (!isWelcomeShown) {
+    return (
+      <Box
+        sx={{
+          mt: "12%",
+          display: "flex",
+        }}
+      >
+        <WelcomeScreen />
+      </Box>
+    );
+  }
+
+  return <Box sx={{ margin: "10px" }}>TODO</Box>;
 };
 
 export default HomePage;
